Renamed cloudcolors to lakefront colors where appropriate.
