+++ src/emotion.d.ts
@@ -1,6 +1,6 @@
 import '@emotion/react';
-import colors from 'core/styles/cloudColors';
-import borders from 'core/styles/borders';
+import colors from 'src/styles/lakefrontColors';
+import borders from 'src/styles/borders';
 
 type ColorsType = typeof colors;
 type BordersType = typeof borders;
