+++ src/index.ts
@@ -29,7 +29,7 @@ export { default as TextArea } from'./TextArea/TextArea';
 
 export { default as Toggle, ToggleOption } from './Toggle/Toggle';
 
-export { default as colors } from './styles/cloudColors';
+export { default as colors } from './styles/lakefrontColors';
 
 export { default as theme } from './styles/theme';
