+++ src/styles/borders.ts
@@ -1,4 +1,4 @@
-import colors from './cloudColors';
+import colors from './lakefrontColors';
 
 export default {
     primary: `1px solid ${colors.selago}`
